requirements.txt
# Core database and Git dependencies
psycopg[binary]==3.2.3
pgvector==0.3.6
sqlmodel==0.0.25
GitPython==3.1.45
python-dotenv==1.0.1
numpy==1.26.4
pyyaml==6.0.2

# GitHub API integration
PyGithub==2.5.0

# Instala el paquete MCP y su grupo de dependencias opcionales [server]
mcp==1.14.1
fastapi-mcp==0.4.0

# LangChain + Ollama analysis dependencies
langchain==0.3.7
langchain-core>=0.3.47
langchain_community==0.3.7
langchain-ollama==0.3.0

# Token estimation
tiktoken==0.11.0
